MA-22090 Refine buffer handle operation in hwcomposer3
As imx95 use arm DDK gralloc to allocate memory and the buffer handle is
different from our buffer handle for imx8 series, evk_95 cannot use hwcomposer3
directly. Here refine the buffer handle operation in hwc3 to make it possible
to make it compatible between imx8 and imx9.
Here just for imx8 part and device composer related is not changed because
2d composition in imx95 still need to study.

Change-Id: I606eebf8c7f501cd70ea1c1288f35fa98da1e083
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/23955
Reviewed-by: Jessie Hao <juan.hao@nxp.com>