MA-22116 hwc3: clear DrmBuffer cache of plane buffer when stop using Overlay
Increase the DrmBuffer cache size for plane buffer(used in Overlay) to make it
cover all the buffers used in video playback.
Clear the DrmBuffer cache when Overlay is not used for more than 2 seconds. So
the buffer of video frame (through Overlay) can be freed when not use for 2s.

Change-Id: I4daf9cf832fd56324be517ca2ab3195ccb335622
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/23954
Reviewed-by: Jessie Hao <juan.hao@nxp.com>