MA-22120 Add synchronization mutex for load firmware
Mutex locks can fix the case in GTS:
com.google.android.wvts.MediaDrmParameterizedTests#testResourceRatingTierNumKeys[L1]

The reason for the failure is that this case needs to run two secure videos
at the same time, and then they load the firmware into the VPU memory at
the same time. The first load has not ended yet, and the second load
will also modify the memory corresponding to the firmware, causing the
firmware to be loaded failed. The normal logic is that after the first
load is successful, the VPU register will be updated to notify the
firmware that the load has been successful, so that there is no need to
load the second load. So if we play a secure video first, this case
will pass. Add a mutex to make the first load as successful as possible
before performing the second load.

Change-Id: I9edcd73b5778790d53f0b5a365cf4d566370665d
Signed-off-by: He Zhai <zhai.he@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/23968
Reviewed-by: Elven Wang <elven.wang@nxp.com>