MA-22227 fix tflite preview stuck issue on 8mq
Issue occurs when opening tflite preview on
the secondary screen during multi-screen display.

In this case, due to the layout of the apk to
adapt the display, the preview turned to horizontal,
camera hal got the 240p resolution to ConfigAndStart
the streams, and we change the 240p to 480p on 8mq,
and due to the logical flaw, the resolution of src and
dst always be unequal, causing the stream always be
skipped.

Fix the issue by treating the width or height of
src and dst as the same condition when skipping the stream.

Change-Id: Ibf6a538d73cb8619049e97ed25cc6082bcf2f986
Signed-off-by: Anle Pan <anle.pan@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/24179
Reviewed-by: Hui Fang <hui.fang@nxp.com>