MA-22274 audio: Support card sequence changing for wm8962
Evk_95 supports both cs42888 and wm8962. Just like mek_8q,
we set cs42888 as speaker and wm8962 as headphone (pluggable).
If headphone is not plugged in, cs42888 will be selected
as the output card, provided that the logic sequence of cs42888
is before wm8962.

Remove the tail of "0" in the match rule to support both
wm8960 and wm8962.

Change-Id: I774039203b3656b1af04526e0adec65de4c94697
Signed-off-by: Jindong Yue <jindong.yue@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/24361
Reviewed-by: Hui Fang <hui.fang@nxp.com>