MA-21993 [EVK_8MP]Display: Physical hdmi display becomes 800*600 after hotplug
Before performing a physical HDMI hotplug operation, the display mode is set to
"1920x1080". After the hotplug event, the display mode changes to "800x600".

The root cause of this issue lies in certain devices where, during a hotplug operation,
the handleHotplug() function is executed. Inside this function, obtaining the updated
display configurations requires a delay. To resolve this, this patch introduce a 32ms
delay within the function implementation, ensuring that the connector receives the
most accurate updated display configurations.

Change-Id: Ia07e80f6310509cf63ddb77b6ec8a4ab09b4ded3
Signed-off-by: Dandan Sun <dandan.sun@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/24442
Reviewed-by: Bo Zhang <bo.zhang@nxp.com>