MA-22638-1 Use fence in G2D composition to improve performance
The G2D composition will cost some time and it will affect the present schedule.
The next VSYNC may be missed if the compoisition time is too long and the framework
will skip the next frame if the out fence is not signaled.
Here add fence for G2D composition to sync compositon work and DRM present work.
It will save some time because G2D composition work can be in parallel.

Change-Id: I747f520b9d989547cc520897fbba680998fa8f41
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/25294
Reviewed-by: Jessie Hao <juan.hao@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/25723