MA-22827-2 Specify display port id by bootargs for multi-display
Use bootargs to set the display port id of each display device like "setenv
append_bootargs androidboot.display_port=HDMI-A-1:2,HDMI-A-2:6". It sets display
port as 2 for HDMI-A-1, 6 for HDMI-A-2.
The "HDMI-A-1" is the name of connector. It can be got from "/sys/class/drm/"
and there is folder like "card0-HDMI-A-1".
Generally the display with minimum value of port id will regard as primary display
and the display id(hwc id) is 0 in framework. The secondary display id(hwc id)
will be the same as display port id.

By the way, the display port id of mxsfb-drm(lcdif in imx8mq) must set the 0x40
bit or display cannot work.

Change-Id: Ie287d0c95a614e67c550888fb54a65a7bc6d91fa
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/25843
Reviewed-by: Jessie Hao <juan.hao@nxp.com>