MA-22412-1 Refine Gralloc HAL for legacy IMX platforms
As the original gralloc handle structure cannot meet the requirement of aidl
version Allocator and stable-c version mapper(usage need to be 64bit), the
change for handle structure is must. A lot of unused code in display/display
folder and libfsldisplay.so make gralloc ugly and be hard to maintain.

Here drop the libfsldisplay.so and remove unused gralloc_v4 code from gralloc
folder. Refine the gralloc logic for dma_buf allocator and VSI gralloc to make
the gralloc framework more clear. Rename some files name and remove all unused
definition and unused funciton to make it easy to maintain.

Change-Id: I32cbd28032b92a2f5bfac00752c35e16dde8d3e3
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/26096
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>