MA-22904 camera: Validate stream config before configuring pipeline
This is required as per libcamera documentaion:

https://libcamera.org/api-html/classlibcamera_1_1CameraConfiguration.html

"Once the camera configuration satisfies the application, it shall be
validated by a call to validate(). The validation implements "try"
semantics: it adjusts invalid configurations to the closest achievable
parameters instead of rejecting them completely. Applications then decide
whether to accept the modified configuration, or try again with a
different set of parameters. Once the configuration is valid, it is
passed to Camera::configure()."

Change-Id: I22fdecd9f5dc9082ec198ef55ae0f0a9abb30568
Signed-off-by: Antoine Bouyer <antoine.bouyer@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/25991
Reviewed-by: Hui Fang <hui.fang@nxp.com>