MA-23175 audio: Put cs42888 to the beginning of the card list
On mek_8q rev E board, cs42888 is probed after wm896x.
The audio hal selects the first avaiable device to be the
input device.
Move cs42888 to the beginning of the card list so that it
can be set as the input device once it is connected.
If audio card is not connected, wm896x is the input device.

Change-Id: I7c66729aaa35742fefcd794825f8499a1faff044
Signed-off-by: Jindong Yue <jindong.yue@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/26632
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>