MA-22932 Add the function of partially deleting cache buffer
When there are multiple overlays and multiple videos are played, all
planedrmbuffers will be released together after all videos are stopped.
This will cause insufficient secure memory when switching videos, so
this function is added when the number of overlays is reduced , retain
some buffers that ara at beginning in the cache.

Change-Id: Ie2877ad394d37f120eeb403edf9b7e7aec4859ec
Signed-off-by: He Zhai <zhai.he@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/26052
Reviewed-by: Bo Zhang <bo.zhang@nxp.com>