audio: Add warning log for unmatched mixer values
Compare the mixer control values set by the user with the values
retrieved from the hardware, and log a warning if they do not match.

This check is meaningful because of the integer division involved in
converting between percentage and hardware values in tinyalsa, which
can lead to small discrepancies due to rounding.

Bug: 375030900
Test: build locally
Change-Id: If396a5cedc768f2bab5db055b5cf875143e3c23b
Signed-off-by: Jindong Yue <jindong.yue@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/26178
Reviewed-by: Hui Fang <hui.fang@nxp.com>