MA-23141 audio: Fix click sound conflict with direct stream
With click sound, the direct stream fails to open the same
card. The default settings of ro.audio.flinger_standbytime_ms
is 3000, which means that the click sound blocks the direct
stream for about 3 seconds.

To fix this, we give the direct stream high priority to use
the pcm device. More specifically, lock the card when starting
the direct stream and release it when the primary stream detects
this lock during transfer, and let the primary stream sleep for a while.

Change-Id: I06dd7a96413f5f642cc35a7856f3a131290318bf
Signed-off-by: Jindong Yue <jindong.yue@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/26535
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>