MA-23330 audio: Reject creating multi-channel stream for non cs42xxx speaker device
On mek_8q and evk_95, the cs42xxx card supports multi-channel
direct output, it acts as a speaker device together with
the wm896x card. Some CTS cases fail to open a
multi-channel stream when the cs42xxx card is not present.

Fix it by rejecting such cases in the open stream stage.

Test: CtsMediaAudioTestCases
  android.media.audio.cts.AudioTrackTest#testPlayStaticShortArray
Change-Id: Ieeeb3e74a2907be009dde4f0bf0394ef083d9004
Signed-off-by: Jindong Yue <jindong.yue@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/27234
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>