MA-23364 audio: Stop retrying start after several times
On mek_8q and evk_95, the cs42xxx card supports multi-channel
direct output, it acts as a speaker device together with
the wm896x card.  Some CTS cases tries to play multi-channel
audio and verify the consumed frames. If the cs42xxx card
is not present, the case fails to open the stream and keep
trying in the transfer function. It takes about 20ms for one
tryStart in a transfer loop, this will make the final frames
not same as expected.

So this patch sets the maximum retry count for starting stream.
After that, the thread will sleep in transfer instead of keep retrying.

Test: CtsMediaAudioTestCases
  android.media.audio.cts.AudioTrackTest#testPlayStaticShortArray
Change-Id: I84979d79c3cc35e714cba7aa220b098bfdbcf649
Signed-off-by: Jindong Yue <jindong.yue@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/27290
Reviewed-by: Hui Fang <hui.fang@nxp.com>